requirements.txt
autopep8==2.0.4
blinker==1.7.0
click==8.1.7
Flask==3.0.0
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.2.0
packaging==23.2
protobuf==4.21.12
pycodestyle==2.11.1
tomli==2.0.1
Werkzeug==3.0.1
